// src/common/interfaces/player.ts
export type Player = {
  id: number;
  competition_id: number;
  full_name: string;
  first_name: string;
  last_name: string;
  known_as: string;
  shorthand: string;
  age: number;
  height: number;
  weight: number;
  league: string;
  league_type: string;
  season: string;
  starting_year: number;
  ending_year: number;
  url: string;
  club_team_id: number;
  club_team_2_id: number;
  national_team_id: number;
  position: string;
  minutes_played_overall: number;
  minutes_played_home: number;
  minutes_played_away: number;
  birthday: number;
  nationality: string;
  continent: string;
  appearances_overall: number;
  appearances_home: number;
  appearances_away: number;
  goals_overall: number;
  goals_home: number;
  goals_away: number;
  clean_sheets_home: number;
  clean_sheets_away: number;
  clean_sheets_overall: number;
  conceded_overall: number;
  conceded_home: number;
  conceded_away: number;
  assists_overall: number;
  assists_home: number;
  assists_away: number;
  penalty_goals: number;
  penalty_misses: number;
  penalty_success: number;
  goals_involved_per_90_overall: number;
  assists_per_90_overall: number;
  goals_per_90_overall: number;
  clean_sheets_per_overall: number;
  goals_per_90_home: number;
  goals_per_90_away: number;
  min_per_goal_overall: number;
  conceded_per_90_overall: number;
  min_per_conceded_overall: number;
  cards_overall: number;
  yellow_cards_overall: number;
  red_cards_overall: number;
  min_per_match: number;
  min_per_card_overall: number;
  min_per_assist_overall: number;
  cards_per_90_overall: number;
  rank_in_league_top_attackers: number;
  rank_in_league_top_midfielders: number;
  rank_in_league_top_defenders: number;
  rank_in_club_top_scorer: number;
  last_match_timestamp: number;
};

export type PlayerIndividualType = {
  id: string;
  competition_id: string;
  full_name: string;
  first_name: string;
  last_name: string;
  known_as: string;
  shorthand: string;
  age: string;
  league: string;
  league_type: string;
  season: string;
  starting_year: string;
  ending_year: string;
  national_team_id: string;
  url: string;
  club_team_id: string;
  club_team_2_id: string;
  position: string;
  minutes_played_overall: string;
  minutes_played_home: string;
  minutes_played_away: string;
  birthday: string;
  nationality: string;
  continent: string;
};